Move the SIGINT handler to separeate file.
Ignore __pycache__.
